refactor: 코드리뷰 반영 - Command 분리 및 Mapper 패턴 적용
- Command 객체를 개별 파일로 분리
- Entity 변환 로직을 Mapper(extension function)로 분리
