+++ src/main/kotlin/com/flab/license/service/plan/command/CreatePlanCommand.kt
@@ -1,16 +1,9 @@
 package com.flab.license.service.plan.command
 
 import com.flab.license.domain.plan.PlanCode
-import com.flab.license.domain.plan.PlanId
 
 data class CreatePlanCommand(
     val planCode: PlanCode,
-    val maxSeats: Int,
-    val monthlyTokenLimit: Long
-)
-
-data class UpdatePlanCommand(
-    val planId: PlanId,
     val maxSeats: Int,
     val monthlyTokenLimit: Long
 )
