+++ src/main/kotlin/com/flab/license/service/plan/command/UpdatePlanCommand.kt
@@ -1,0 +1,9 @@
+package com.flab.license.service.plan.command
+
+import com.flab.license.domain.plan.PlanId
+
+data class UpdatePlanCommand(
+    val planId: PlanId,
+    val maxSeats: Int,
+    val monthlyTokenLimit: Long
+)
